% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A_DATA_2}
\alias{A_DATA_2}
\title{Analysis Data}
\format{
A data frame with 101316 rows and 133 variables:
\describe{
  \item{SEQN}{respondent identifier}
  \item{DIABETES}{Diabetic status of the respondent, 1 for "Yes", 0 for "No", NA otherwise}
}
}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/Default.aspx}
}
\usage{
A_DATA_2
}
\description{
An extract of the NHANES data.
}
\keyword{datasets}
