% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{add_two_numbers}
\alias{add_two_numbers}
\title{Add numbers}
\usage{
add_two_numbers(x, y)
}
\arguments{
\item{x}{a numeric vector that will be added with y}

\item{y}{a numeric vector that will be added with x}
}
\value{
a numeric vector that is the sum of x and y
}
\description{
Add numbers
}
\examples{
subtract_two_numbers(2, 2)

subtract_two_numbers(c(2, 3, 5), c(2, 4, 9))

}
