% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_two_nums.R
\name{add_two_nums}
\alias{add_two_nums}
\title{add two numbers}
\usage{
add_two_nums(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{a number}
}
\value{
the sum of x and y
}
\description{
A sample function that adds two inputs
}
\examples{
\dontrun{
add_two_nums(3,4)
}
}
