% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FEATURE_TYPE}
\alias{FEATURE_TYPE}
\title{Feature Type}
\format{
list of 3
\describe{
  \item{Not_features}{identifier, targets, ect}
  \item{categorical_features}{list of categorical features }
  \item{numeric_features}{list of numerical features}
}
}
\usage{
FEATURE_TYPE
}
\description{
A list of categorized features
}
\keyword{datasets}
