% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_model_scorer.R
\name{logit_model_scorer}
\alias{logit_model_scorer}
\title{logit model scorer}
\usage{
logit_model_scorer(my_model, my_data, target, level)
}
\arguments{
\item{my_model}{a logistic glm model}

\item{my_data}{data to be scored}

\item{target}{outcome in glm model}

\item{level}{the level for which the glm should be set}
}
\value{
my_data will be returned along with columns: model, probs, pred
}
\description{
A wrapper function that performs scoring on logistic glm model
}
\examples{
\dontrun{
logit_model_scorer(logit_glm, A_DATA_2.test, DIABETES_factor, 1)
}

}
