% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic.R
\name{subtract_two_numbers}
\alias{subtract_two_numbers}
\title{Subtract numbers}
\usage{
subtract_two_numbers(x, y)
}
\arguments{
\item{x}{a numeric vector that y will be subtracted from}

\item{y}{a numeric vector to be subtracted from x}
}
\value{
a numeric vector that is the difference between x and y
}
\description{
Subtract numbers
}
\examples{
subtract_two_numbers(2, 2)

subtract_two_numbers(c(2, 3, 5), c(2, 4, 9))

}
