% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_t_ks_test.R
\name{wrapper.t_ks_test}
\alias{wrapper.t_ks_test}
\title{t-test / ks-test wrapper}
\usage{
wrapper.t_ks_test(
  df,
  factor,
  factor_level,
  continuous_feature,
  verbose = FALSE
)
}
\arguments{
\item{df}{a tibble - data-frame or connection}

\item{factor}{a categorical variable in df ; note in SQLlite there is no factor variable but the user thinks of this variable as a factor}

\item{factor_level}{\itemize{
\item sets the \strong{"first"} class in a \strong{one-versus-rest} \code{t.test} and \code{ks.test} analysis
}}

\item{verbose}{was useful in creating the function to identify variables causing errors, and create output for those types of cases, set to FALSE by default}

\item{continuous_feature`}{represented by a character string of the feature name to perform \code{t.test} and \code{ks.test} on}
}
\value{
A row-tibble with columns: Feature, mean_diff_est, ttest.pvalue, mean_Target, mean_Control, kstest.pvalue, N_Target, N_Control, conf.low, conf.high, statistic, sd_Target, sd_Control, parameter, method, alternative
}
\description{
A wrapper function that performs a one-versus the rest t-test & ks-test
}
\examples{
\dontrun{
wrapper.t_ks_test(df=A_DATA_TBL_2 ,
                 factor = DIABETES ,
                 factor_level = 1,
                 'Age')
}

}
